backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
openai==1.3.5
email-validator==2.1.0
phonenumbers==8.13.25
pyotp==2.9.0
qrcode[pil]==7.4.2
authlib==1.2.1
httpx==0.25.2
python-decouple==3.8
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
mediapipe==0.10.13
opencv-python==4.8.1.78
numpy==1.24.3
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
reportlab==4.0.7

